Added fixes so that mininush will build on Linux.
The Linux implementation of Nu appears to be broken at the moment.
These changes at least get you to the point where mininush can be
built, albeit with many warnings. Running
./mininush tools/nuke
fails though, because of the error

Uncatched Objective-C exception:
exceptionClass <class NSException>
Reason: unable to load nu
UserInfo: no userinfo

Please see the new notes file notes/LINUXINSTALL for a full
explanation of what needs to be done to get to this stage
using Ubuntu 8.04 running in a VM on Mac OS 10.5.3.