Added an optional @framework_install_path ivar to framework-tasks in nuke.
If set, @framework_install_path will be prepended to the @framework_executable_name
in order to generate the Framework path for the -install_name flag that
gets passed to the linker. Setting the @framework_install_path in the Nukefile
is equivalent to setting the INSTALL_PATH or 'Installation Directory' (Xcode >= 3.0)
key for the Deployment Target in Xcode.  The most common setting is:

(set @framework_install_path "@executable_path/../Frameworks")

This is convenient for building Frameworks that will be installed inside a bundle.

Leave @framework_install_path unset to get the default behavior with the
install_name set relative to the root of the framework directory.