Changes to allow Nu to run with GNUstep-base on Linux.
Tested on Debian "Etch" running under VMware.

NuFound (aka libFoundation) support is now dropped,
since GNUstep is actively-maintained and contains
important features missing in libFoundation, notably
UTF-8 support.

Because the Linux build is running on the GNU
Objective-C runtime, the "throw" operator is not
available, and because the Nu "match" operator
is currently a heavy user of "throw", use of
the "match" operator should be avoided on Linux.