Automatic install and test tasks for frameworks, visibility improvements, bug fix.
We now automatically add "install" and "test" tasks for frameworks.
Set @framework_extra_install to a block to add any additional installation actions.
I also fixed an error in dependency result calculation: Info.plist creation
for frameworks was not returning zero, causing the framework builds to end
early. As part of this, the -v (verbose) option now displays more information
and causes action blocks to be printed when the -T (show tasks) option is given.