new tricks for setting NSMutableDictionary entries.
I have been growing tired of writing setObject:forKey: whenever I need
to set values in dictionaries.

For a long time now, we've been able to access dictionary entries by
simply providing keys as virtual methods that are implemented with
[NSDictionary(Nu) handleUnknownMessage:withContext:]

% (set d (dict a:12 b:23 c:34))
<NSCFDictionary:3805ec0>
% (d "a")
12

Originally, I chose to require "a" to be a string (and not a symbol)
to allow this same access mechanism to be used with keys that are
represented by variables.

% (set x (array "a" "b" "c"))
<NSCFArray:38064a0>
% (x each:(do (i) (puts (d i))))
12
23
34
<NSCFArray:38064a0>

Of course this is all sugar; we could have made the same accesses
using [NSDictionary objectForKey:], but this lets us type less and is
fairly intuitive.

Along those same lines, I have a change in my local git repository
that allows me to set dictionary contents  with a similar shorthand:

% (d a:56)
% (d "a")
56

Also, since symbols that end with a semicolon are already treated
specially (I call them "labels" and they evaluate to themselves), I
decided to make this possible also:

% (d a:)
56

All of this is implemented in a rewritten version of [NSDictionary(Nu)
handleUnknownMessage:withContext:]

Also, drawing code from [NSDictionary(Nu) dictionaryWithList:], it was
easy to go ahead and support the following, which is similar to the
existing "dict" operator:

% (d a:78 d:89 e:90)
% (puts (d description))
{
   a = 78;
   b = 23;
   c = 34;
   d = 89;
   e = 90;
}
()

As a seemingly-reasonable side effect, this also works:

% (d a:11 b:22 c:33 a:)
11

A couple of notes:

1. (d a:11 b:22 c:33) is equivalent to (d a: 11 b: 22 c: 33). The Nu
parser always ends symbols on semicolons.

2. This would generate exceptions if you tried to modify immutable
dictionaries.

% (set d (NSDictionary dictionaryWithObject:123 forKey:"a"))
<NSCFDictionary:3807470>
% (d "a")
123
% (d a:)
123
% (d a:456)
NSInternalInconsistencyException: *** -[NSCFDictionary
nuSetObject:forKey:]: mutating method sent to immutable object