Add XIB file compilation tasks to nuke.
When one or more XIB files are specified in a Nukefile, nuke will now
compile them with ibtool, the same executable that Xcode uses to compile
XIB files.

If you put your XIB files in a directory called "resources", the following
Nukefile declaration will cause them to be compiled (to NIB files) and
copied into your application or framework's Resources directory:

    (set @xib_files  (filelist "^resources/.*\.xib$"))

The compilation is triggered by the nuke "compilation-tasks" macro,
and the "application-tasks" and "framework-tasks" macros cause
the compiled file to be copied into the application or framework bundle,
respectively.