Added a small hack to the NSMethodSignature workaround we're using to side-step a bug in Apple's objective C runtime Added a small hack to the NSMethodSignature workaround we're using to side-step a bug in Apple's objective C runtime method_getNumberOfArguments() and method_getArgumentType() functions.
The method_getNumberOfArguments() and method_getArgumentType() functions count block objects twice
(their type signature is "@?" and it counts both the '@' and the '?') so we use -[NSMethodSignature getNumberOfArguments]
which correctly reports the number of arguments even if some are block objects.  However, getNumberOfArguments returns zero
if a method has no arguments, while method_getNumberOfArguments() returns 2 (for the implicit arguments cmd and self).
This commit adds a hack that falls back to method_getNumberOfArguments() and method_getArgumentType() if
-[NSMethodSignature getNumberOfArguments] return zero.

This allows us to use pass blocks into objective c methods, and it doesn't break 'nuke test'.