operator cleanup
The (ivars) and (ivar-accessor) operators now do nothing and should
be removed. Instance variables are now implicitly created and
stored using associated objects, so there is no overhead for their
potential use.

The (cmethod) and (imethod) operators are still available but with
a deprecation warning in the log. They should be replaced by (+) and
(-), which are more consistent with Objective-C usage.

The (def) operator is gone. It was carried over from Ruby and in
practice, the equivalent (function) is preferred.