Check whether the receiver specifies forwardingTarget for the given selector.
After exhausting all other possible outcomes, add a call to -forwardingTargetForSelector: to check whether there is another object that should receive the message.  If so, then send try sending the message to that target. This allows classes defined in Nu to take advantage of "fast" message forwarding by implementing -forwardingTargetForSelector: