improved handling of init* methods that don't return self.
NSXMLDocument is one of many classes whose init methods return
something other than self. Others include NSArray and NSDictionary,
which return a special "placeholder" subclass instance from alloc.
NSXMLDocument behaves similarly but without an explicit placeholder
subclass.

This change should eliminate a crash that occurred when placeholders
were overreleased.

In past versions of Nu, we looked for known explicit placeholders
and handled them specially, but this was removed in Lion when the
NSArray and NSDictionary placeholders subclasses appeared to to
ignore release messages. However, this was not true for
NSXMLDocument, which implicitly uses the placeholder pattern and
does not have a placeholder subclass.

Now we watch methods whose names begin with "init" to see if they
return their original target or not, which allows us to properly
handle the placeholder pattern whether it is used from an explicit
placeholder subclass or not.