New "nubake" feature: easier baking into classes.
nubake accepts three new options:
	--method <methodname>
	--class <classname>
	--category <categoryname>

When these are provided, nubake produces a source file that adds
the named category to the named class. The category adds a method
that evaluates the baked Nu code and returns the result. The
purpose of this is to bake in code that can be evaluated from Nu.
The code is evaluated in the main Nu parser's context so that
macros and other objects can be made available to subsequent
Nu code.

Example:
% nubake server.nu --method macros --category Baked --class MyServer --output MyServer+Baked.m