Add support for embedded classes and ids to NuMarkupOperator
I've been generating a lot of HTML with Nu and decided to
take this simple extension from the Hiccup Clojure library.

Now when we specify a markup element with the (&element)
automatic operator, we can add classes and id values by
including "." and "#" in the element name.

Examples:

% (puts (&div))
<div></div>

% (puts (&div.x))
<div class="x"></div>

% (puts (&div#main))
<div id="main"></div>

% (puts (&div#main.x1.x2.x3))
<div id="main" class="x1" class="x2" class="x3"></div>

Beware that this means that the '#' can now be included in symbols,
so if you have code like "this#", you will be unpleasantly surprised
by this change.